refactor: split render.ts and hooks.ts into focused submodules
- Split src/hooks.ts into src/hooks/ with one file per hook:
  symbols.ts, use-state.ts, use-ref.ts, use-id.ts, use-memo.ts,
  use-resolve.ts, use-effect.ts, use-render.ts, use-ui-patch.ts, index.ts
- Split src/render.ts into src/render/ by concern:
  types.ts, state.ts, events.ts, helpers.ts, props.ts,
  hooks-runtime.ts, patch.ts, mount.ts, reconciler.ts, index.ts
- Use a shared renderState object to share mutable state across ES modules
- All 132 tests pass; build and formatting clean
